requirements.txt
google-api-python-client==2.115.0
anthropic==0.40.0
python-dotenv==1.0.1
